Fix crud controller of Contact

+++ src/Controller/ContactController.php
@@ -10,26 +10,31 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\Security\Core\User\UserInterface;
 
 #[Route('/contact')]
 class ContactController extends AbstractController
 {
     #[Route('/', name: 'app_contact_index', methods: ['GET'])]
-    public function index(ContactRepository $contactRepository): Response
+    public function index(ContactRepository $contactRepository, UserInterface $user): Response
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
         return $this->render('contact/index.html.twig', [
-            'contacts' => $contactRepository->findAll(),
+            'contacts' => $contactRepository->findBy(['username' => $user]),
         ]);
     }
 
     #[Route('/new', name: 'app_contact_new', methods: ['GET', 'POST'])]
-    public function new(Request $request, EntityManagerInterface $entityManager): Response
+    public function new(Request $request, EntityManagerInterface $entityManager, UserInterface $user): Response
     {
         $contact = new Contact();
         $form = $this->createForm(ContactType::class, $contact);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+            $contact->setUsername($user);
+            $contact->setCreatedAt(new \DateTimeImmutable());
+            $contact->setUpdatedAt(new \DateTimeImmutable());
             $entityManager->persist($contact);
             $entityManager->flush();
 
@@ -43,20 +48,28 @@ class ContactController extends AbstractController
     }
 
     #[Route('/{id}', name: 'app_contact_show', methods: ['GET'])]
-    public function show(Contact $contact): Response
+    public function show(Contact $contact, UserInterface $user): Response
     {
+        if ($user !== $contact->getUsername()){
+            return $this->redirectToRoute('app_contact_index');
+        }
         return $this->render('contact/show.html.twig', [
             'contact' => $contact,
         ]);
     }
 
     #[Route('/{id}/edit', name: 'app_contact_edit', methods: ['GET', 'POST'])]
-    public function edit(Request $request, Contact $contact, EntityManagerInterface $entityManager): Response
+    public function edit(Request $request, Contact $contact, EntityManagerInterface $entityManager, UserInterface $user): Response
     {
+
+        if ($user !== $contact->getUsername()){
+            return $this->redirectToRoute('app_contact_index');
+        }
         $form = $this->createForm(ContactType::class, $contact);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+            $contact->setUpdatedAt(new \DateTimeImmutable());
             $entityManager->flush();
 
             return $this->redirectToRoute('app_contact_index', [], Response::HTTP_SEE_OTHER);
@@ -69,8 +82,11 @@ class ContactController extends AbstractController
     }
 
     #[Route('/{id}', name: 'app_contact_delete', methods: ['POST'])]
-    public function delete(Request $request, Contact $contact, EntityManagerInterface $entityManager): Response
+    public function delete(Request $request, Contact $contact, EntityManagerInterface $entityManager, UserInterface $user): Response
     {
+        if ($user !== $contact->getUsername()){
+            return $this->redirectToRoute('app_contact_index');
+        }
         if ($this->isCsrfTokenValid('delete'.$contact->getId(), $request->request->get('_token'))) {
             $entityManager->remove($contact);
             $entityManager->flush();
